+++ app/views/employment/EmployerInformationView.scala.html
@@ -19,9 +19,9 @@
 @import views.html.templates.helpers.NotificationBanner
 @import utils.InYearUtil
 @import utils.ViewUtils.summaryListRow
-@import controllers.employment.EmployerInformationRow
 @import controllers.employment.routes._
 @import controllers.studentLoans.routes._
+@import viewmodels.employment._
 
 @this(
         layout: Layout,
